+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added inspect command to `cartesi-rollups-cli`
 - Added increase-time command to `cartesi-rollups-cli`
 - Added instructions on how to run the node with Docker.
+- Added validate command to `cartesi-rollups-cli`
 
 ### Changed